I've been messing with the config, and I have to stop working now
Its got some retarded stuff in it, but dont worry.
Since each line is a string, it prints out in the config with ' ' on
both sides. Which is incredibly annoying to me, and i figured out that
you can get rid of them by removing the spaces in the string, which
obviously doesnt work for all of them, so Im kinda stuck. I might be
reformatting some stuff, such as the levels.1 to a Level_1, to get rid
of the quotes on the 1, but this is really bothering me, lol.

Basically I got getting and setting of the effects working (as far as I
know, since there arent actually classes yet, so it throws errors), but
then I broke it by reformatting everything. So ill be working on this
more later, lol.