Lots of config work...
looks like, unless we wanna switch entirely to hyphens or something,
the quotes in the config are unnegotiable because without them it wont
read in correctly. Which is kinda a bummer, but, we'll deal with that
later. For now, the config reads the distances, levels, messages,
effects, and beacon item point values correctly and stores them in
arrays/etc. for more efficient access. I havent added anything about
the mirages or mob spawns yet.

I also added 2 classes for TorchConsume and LifeDrain, to test my
config reading, and an Effect class, which they will all extend. So
far, all 3 are empty.