Fixed your persistence errors ;) Also, config is 100% functional
There's a shitload of debug code in the config set method, you can turn
it off by commenting out the log method in the main class.

The config now completely works for everything except mirages
attributes, since Im not sure what's going to actually be there yet.
But right now, it reads everything correctly and stores everything in
the arrays at the top and all of my get methods are beautiful
one-liners: couldn't possibly be cleaner or more efficient :)

Added a package for mirages and a couple more empty classes.

Updated gitignore.

Also, I accidentally added the beacon persister to the gitignore. I
removed it, but right now in my changes it still shows a deleted class,
and then a new class by that name, with everything still in it. It
shouldnt cause any issues, but if it does, thats why... lol